chore(mcp_store): drop consumer-less JSON-RPC endpoint and facade wrappers
